Move precaching fns to profile where the cache is
